+++ src/L10NSharp/XLiffUtils/XLiffBody.cs
@@ -14,8 +14,11 @@
 // </remarks>
 // ---------------------------------------------------------------------------------------------
 using System;
+using System.Collections;
+using System.Collections.Concurrent;
 using System.Linq;
 using System.Collections.Generic;
+using System.Threading;
 using System.Xml.Serialization;
 using L10NSharp.XLiffUtils;
 
@@ -29,34 +32,63 @@ namespace L10NSharp.XLiffUtils
 		// This is used when translation unit IDs are not found in the file (which seems to be
 		// the case with Lingobit XLiff files).
 		private int _transUnitId;
-		private bool _idsVerified;
-		private List<XLiffTransUnit> _transUnits = new List<XLiffTransUnit>();
+		static SpinLock _transUnitIdLock;
+
+		private ConcurrentDictionary<string, XLiffTransUnit> _transUnitDict =
+			new ConcurrentDictionary<string, XLiffTransUnit>();
+
+		private object mutex = new object(); // lock for accessing non-concurrent variables.
 		private int _translatedCount = -1;
 		private int _approvedCount = -1;
 
+		public class ListWrapper : IEnumerable<XLiffTransUnit>
+		{
+			private IEnumerable<XLiffTransUnit> _list;
+			private XLiffBody _body;
+
+			public ListWrapper(IEnumerable<XLiffTransUnit> startWith, XLiffBody body)
+			{
+				_list = startWith;
+				_body = body;
+			}
+			public IEnumerator<XLiffTransUnit> GetEnumerator()
+			{
+				return _list.GetEnumerator();
+			}
+
+			IEnumerator IEnumerable.GetEnumerator()
+			{
+				return GetEnumerator();
+			}
+
+			public void Add(XLiffTransUnit item)
+			{
+				_body.AddTransUnitRaw(item);
+			}
+		}
+
 		#region Properties
-		/// ------------------------------------------------------------------------------------
+
 		/// <summary>
-		/// Gets the list of translation units in the file.
+		/// This property exists solely to support serializing and deserializing an XliffBody in a
+		/// backwards-compatible way. The serialization code assumes it can get the object and use
+		/// Add to put things into it when deserializing as well as running the enumeration to get
+		/// the things to serialize. The ListWrapper implements just those necessary functions.
+		/// This property must be public for the serializer to work, but is not intended for use
+		/// by other clients.
 		/// </summary>
-		/// ------------------------------------------------------------------------------------
 		[XmlElement("trans-unit")]
-		public List<XLiffTransUnit> TransUnits
+		public ListWrapper TransUnitsForXml
 		{
 			get
 			{
-				if (!_idsVerified && _transUnits != null && _transUnits.Count > 0)
-				{
-					foreach (var tu in _transUnits.Where(tu => string.IsNullOrEmpty(tu.Id)))
-						tu.Id = (++_transUnitId).ToString();
-
-					_idsVerified = true;
-				}
-
-				return _transUnits;
+				var result = TransUnitsUnordered.ToList();
+				result.Sort(XLiffLocalizedStringCache.TuComparer);
+				return new ListWrapper(result, this);
 			}
-			set { _transUnits = value; }
 		}
+
+		[XmlIgnore] public IEnumerable<XLiffTransUnit> TransUnitsUnordered => _transUnitDict.Values;
 
 		/// ------------------------------------------------------------------------------------
 		/// <summary>
@@ -66,7 +98,7 @@ namespace L10NSharp.XLiffUtils
 		/// </summary>
 		/// ------------------------------------------------------------------------------------
 		[XmlIgnore]
-		public readonly Dictionary<string, string> TranslationsById = new Dictionary<string, string>();
+		public readonly ConcurrentDictionary<string, string> TranslationsById = new ConcurrentDictionary<string, string>();
 		#endregion
 
 
@@ -78,17 +110,22 @@ namespace L10NSharp.XLiffUtils
 		/// ------------------------------------------------------------------------------------
 		internal XLiffTransUnit GetTransUnitForId(string id)
 		{
-			return _transUnits.FirstOrDefault(tu => tu.Id == id);
+			_transUnitDict.TryGetValue(id, out XLiffTransUnit result);
+			return result;
 		}
 
 		/// <summary>
 		/// When all but the last part of the id changed, this can help reunite things
 		/// </summary>
-		internal XLiffTransUnit GetTransUnitForOrphan(XLiffTransUnit orphan)
+		internal XLiffTransUnit GetTransUnitForOrphan(XLiffTransUnit orphan, XLiffBody source)
 		{
 			var terminalIdToMatch = XLiffLocalizedStringCache.GetTerminalIdPart(orphan.Id);
 			var defaultTextToMatch = GetDefaultVariantValue(orphan);
-			return _transUnits.FirstOrDefault(tu => XLiffLocalizedStringCache.GetTerminalIdPart(tu.Id) == terminalIdToMatch && GetDefaultVariantValue(tu) == defaultTextToMatch);
+			return TransUnitsUnordered.FirstOrDefault(tu =>
+				XLiffLocalizedStringCache.GetTerminalIdPart(tu.Id) ==
+				terminalIdToMatch // require last part of ID to match
+				&& GetDefaultVariantValue(tu) == defaultTextToMatch // require text to match
+				&& source?.GetTransUnitForId(tu.Id) == null); // and translation does not already have an element for this
 		}
 
 		string GetDefaultVariantValue(XLiffTransUnit tu)
@@ -104,19 +141,44 @@ namespace L10NSharp.XLiffUtils
 		/// <param name="tu">The translation unit.</param>
 		/// <returns>true if the translation unit was successfully added. Otherwise, false.</returns>
 		/// ------------------------------------------------------------------------------------
-		internal bool AddTransUnit(XLiffTransUnit tu)
+		internal bool AddTransUnitRaw(XLiffTransUnit tu)
 		{
 			if (tu == null || tu.IsEmpty)
 				return false;
 
-			if (tu.Id == null)
-				tu.Id = (++_transUnitId).ToString();
+			bool lockTaken = false;
+			string key;
+			try
+			{
+				_transUnitIdLock.Enter(ref lockTaken);
+				// Efficiently lock this very small task so that if we need to modify
+				// the TU, the key that it gets is guaranteed to be the one used to insert
+				// it into the dictionary. This assumes nothing else modifies IDs once they
+				// are in this system: once our locked code has given the TU an ID, any other
+				// thread will see that it is non-empty.
+				key = tu.Id;
+				if (string.IsNullOrEmpty(key))
+				{
+					tu.Id = (System.Threading.Interlocked.Increment(ref _transUnitId)).ToString();
+					key = tu.Id;
+				}
+			}
+			finally
+			{
+				if (lockTaken) _transUnitIdLock.Exit(false);
+			}
 
 			// If a translation unit with the specified id already exists, then quit here.
 			if (GetTransUnitForId(tu.Id) != null)
 				return false;
-
-			_transUnits.Add(tu);
+			_transUnitDict[key] = tu;
+			return true;
+		}
+		public bool AddTransUnit(XLiffTransUnit tu)
+		{
+			if (!AddTransUnitRaw(tu))
+				return false;
+		
 			// If the target exists, store its value in the dictionary lookup.  Otherwise, store
 			// the source value there.
 			if (tu.Target != null && tu.Target.Value != null)
@@ -154,25 +216,15 @@ namespace L10NSharp.XLiffUtils
 		/// Removes the specified translation unit.
 		/// </summary>
 		/// ------------------------------------------------------------------------------------
-		internal void RemoveTransUnit(XLiffTransUnit tu)
+		public void RemoveTransUnit(XLiffTransUnit tu)
 		{
-			if (tu == null)
+			// if the ID is null, it can't be in our dictionary, unless someone
+			// cheated and changed the ID by putting it there after inserting it.
+			if (tu == null || tu.Id == null)
 				return;
 
-			if (_transUnits.Contains(tu))
-			{
-				_transUnits.Remove(tu);
-			}
-			else if (tu.Id != null)
-			{
-				var tmptu = GetTransUnitForId(tu.Id);
-				if (tmptu != null)
-				{
-					_transUnits.Remove(tmptu);
-				}
-			}
-			if (tu.Id != null)
-				TranslationsById.Remove(tu.Id);
+			_transUnitDict.TryRemove(tu.Id, out _);
+			TranslationsById.TryRemove(tu.Id, out _);
 		}
 
 		#endregion
@@ -187,26 +239,30 @@ namespace L10NSharp.XLiffUtils
 		{
 			get
 			{
-				if (_translatedCount < 0)
+				lock (mutex)
 				{
-					_translatedCount = 0;
-					foreach (var tu in TransUnits)
+					if (_translatedCount < 0)
 					{
-						if (tu.Target == null || string.IsNullOrWhiteSpace(tu.Target.Value))
-							continue;
-						if (tu.TranslationStatus == TranslationStatus.Approved ||
-							tu.Target.TargetState == XLiffTransUnitVariant.TranslationState.Translated)
+						_translatedCount = 0;
+						foreach (var tu in TransUnitsUnordered)
 						{
-							++_translatedCount;
-						}
-						else if (tu.Target.Value != tu.Source.Value &&
-							tu.Target.TargetState == XLiffTransUnitVariant.TranslationState.Undefined)
-						{
-							++_translatedCount;
+							if (tu.Target == null || string.IsNullOrWhiteSpace(tu.Target.Value))
+								continue;
+							if (tu.TranslationStatus == TranslationStatus.Approved ||
+							    tu.Target.TargetState == XLiffTransUnitVariant.TranslationState.Translated)
+							{
+								++_translatedCount;
+							}
+							else if (tu.Target.Value != tu.Source.Value &&
+							         tu.Target.TargetState == XLiffTransUnitVariant.TranslationState.Undefined)
+							{
+								++_translatedCount;
+							}
 						}
 					}
+
+					return _translatedCount;
 				}
-				return _translatedCount;
 			}
 		}
 
@@ -220,25 +276,29 @@ namespace L10NSharp.XLiffUtils
 		{
 			get
 			{
-				if (_approvedCount < 0)
+				lock (mutex)
 				{
-					_approvedCount = 0;
-					foreach (var tu in TransUnits)
+					if (_approvedCount < 0)
 					{
-						if (tu.Target == null || string.IsNullOrWhiteSpace(tu.Target.Value))
-							continue;
-						if (tu.TranslationStatus == TranslationStatus.Approved)
-							++_approvedCount;
+						_approvedCount = 0;
+						foreach (var tu in TransUnitsUnordered)
+						{
+							if (tu.Target == null || string.IsNullOrWhiteSpace(tu.Target.Value))
+								continue;
+							if (tu.TranslationStatus == TranslationStatus.Approved)
+								++_approvedCount;
+						}
 					}
+
+					return _approvedCount;
 				}
-				return _approvedCount;
 			}
 		}
 
 		/// <summary>
 		/// Return the total number of strings.
 		/// </summary>
-		internal int StringCount => TransUnits.Count;
+		internal int StringCount => _transUnitDict.Count;
 	}
 
 	#endregion